# Complete Milestone 1.3: Performance Optimization and Caching