--- a/.audit-ci.json
+++ b/.audit-ci.json
@@ -1,0 +1,10 @@
+{
+  "moderate": true,
+  "high": true,
+  "critical": true,
+  "allowlist": [
+  ],
+  "skip-dev": false,
+  "report-type": "full",
+  "output-format": "text"
+}