--- a/config/config.sample.yaml
+++ b/config/config.sample.yaml
@@ -37,6 +37,34 @@ sources:
     enabled: true
     timeout_ms: 5000
     max_retries: 1
+cache:
+  enabled: true
+  strategy: hybrid
+  memory:
+    max_keys: 1000
+    ttl_seconds: 3600
+    check_period_seconds: 600
+  redis:
+    enabled: true
+    url: redis://localhost:6379
+    ttl_seconds: 7200
+    key_prefix: 'pp:cache:'
+    connection_timeout_ms: 5000
+    retry_attempts: 3
+    retry_delay_ms: 1000
+  content_types:
+    runbooks:
+      ttl_seconds: 3600
+      warmup: true
+    procedures:
+      ttl_seconds: 1800
+      warmup: false
+    decision_trees:
+      ttl_seconds: 2400
+      warmup: true
+    knowledge_base:
+      ttl_seconds: 900
+      warmup: false
 embedding:
   enabled: true
   model: sentence-transformers/all-MiniLM-L6-v2