--- a/test-data/generation-summary.json
+++ b/test-data/generation-summary.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2025-07-28T23:59:26.378Z",
+  "generated_at": "2025-07-29T15:25:34.810Z",
   "total_runbooks": 15,
   "total_knowledge_articles": 30,
   "alert_scenarios": 7,